fixed api jsons

+++ Cr-8/src/main/java/com/Cr_8/repositories/BookingRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Cr_8.entities.BookingRequest;
 
-public interface BookingRepo extends JpaRepository<BookingRequest, Integer>{
+public interface BookingRepo extends JpaRepository<BookingRequest, Long>{
 
 }
